Complete
Basic functions complete with API calls and ability to type and have that user displayed using the github users api

Experimenting

playing with different things and the responses that I get to begin to get a deeper understanding of HTTP requests and how to interface with the APIs more comfortably. I want to be able to effortlessly interface with these APIs at a moments notice.